Removing Repository.kt as the main the data access point
If we need to access only local data sources, no access should be available to remote repositories then

By removing Repository.kt and using other local & remote repositories individually,
we make our data classes more testable
