+++ app/src/main/java/com/itranslate/recorder/di/DataModule.kt
@@ -1,7 +1,5 @@
 package com.itranslate.recorder.di
 
-import com.itranslate.recorder.data.Repository
-import com.itranslate.recorder.data.RepositoryImpl
 import com.itranslate.recorder.data.local.sources.LocalRepository
 import com.itranslate.recorder.data.local.sources.LocalRepositoryImpl
 import com.itranslate.recorder.data.remote.sources.RemoteRepository
@@ -11,5 +9,4 @@ import org.koin.dsl.module
 val dataModule = module {
     single<LocalRepository> { LocalRepositoryImpl(get()) }
     single<RemoteRepository> { RemoteRepositoryImpl() }
-    single<Repository> { RepositoryImpl(get(), get()) }
 }
