+++ app/src/main/java/com/itranslate/recorder/ui/fragments/home/HomeViewModel.kt
@@ -2,21 +2,26 @@ package com.itranslate.recorder.ui.fragments.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.paging.Pager
+import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
-import com.itranslate.recorder.data.Repository
 import com.itranslate.recorder.data.local.models.records.Record
+import com.itranslate.recorder.data.local.sources.LocalRepository
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 
 class HomeViewModel(
-    private val repository: Repository
+    private val localRepository: LocalRepository,
+    private val dispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : ViewModel() {
 
     suspend fun insertRecordInDb(record: Record) {
-        return repository.insertRecord(record)
+        return localRepository.insertRecord(record)
     }
 
     private val _recordsFlow: MutableStateFlow<PagingData<Record>> =
@@ -25,7 +30,17 @@ class HomeViewModel(
 
     fun getDbRecords() {
         viewModelScope.launch {
-            repository.getRecords().cachedIn(viewModelScope).collectLatest {
+            val records = localRepository.getRecords()
+            Pager(
+                PagingConfig(
+                    pageSize = 25,
+                    prefetchDistance = 5,
+                    enablePlaceholders = false,
+                    initialLoadSize = 100
+                )
+            ) {
+                records.asPagingSourceFactory(dispatcher).invoke()
+            }.flow.cachedIn(viewModelScope).collectLatest {
                 val previousValue = _recordsFlow.value
                 _recordsFlow.compareAndSet(
                     previousValue,
